"2:45pm. Let me resume. I think I should have enough to start right now, but let me watch some more.
https://youtu.be/pHpTSLZKIC4?t=354

How does he do this. In fact, this is one of the things I need to put the circle where I need it to be, but it hasn't worked for me.

https://youtu.be/pHpTSLZKIC4?t=317
> Snapping dots

https://youtu.be/pHpTSLZKIC4?t=325
> If you want to watching a video on snapping dots and grid, I released a video a few days ago.

Let me track it down.

https://youtu.be/7SF3yCu53Q0
Boxcutter Tutorial for Blender 3.0 - Precision Booleans

I am guessing this is the one.

I actually got the shortcuts pdf.

https://youtu.be/7SF3yCu53Q0?t=52

Oh, I see. I was really missing this kind of capability moving into Blender from Moi. Moi will always show the midpoints, centers and endpoints (amongst others) and snap to them. This is super useful for precision work, which I in fact need in some places. It is really great that hopscutter has this.

3pm. I am having some difficulty. Let me focus more intently. Just how do I turn this feature on. I thought that Ctrl would work.

3:05pm. This is amazing. When I do a boolean cut in edit mode, it does not extrude the box past the other side of the solid. This would be particularly useful for making holes.

This addon really is useful. It would in fact save me a lot of time during modeling. Right now I can easily make cuts, but if I had to do this with booleans by hand, it would be much more tedious.

3:10pm. Oh, snap grid is more useful that I thought it would be at first glance. It would be one thing if it just snapped to grid, but what it in fact does is create a construction plane on the selected face. It is even sized so that the endpoints connect with it...Actually, it seems that was accidental.

But being able to bring up a contruction plane like this would be really useful.

https://youtu.be/7SF3yCu53Q0?t=100

Oh right, the dots respond even if the boolean is not baked in.

3:25pm. https://youtu.be/7SF3yCu53Q0?t=355

Ah, this is not just for dots. Holding alt to make it start in the middle works in general. Nice.

https://youtu.be/7SF3yCu53Q0?t=495

Let me try it out.

https://youtu.be/7SF3yCu53Q0?t=526

Hmmm, this could be useful for those grate patterns in the side. It would be annoying to set up the arrays for those kinds of things. Hmmm, maybe dots would work with arrays?

https://youtu.be/7SF3yCu53Q0?t=638

It is possible to turn off snapping, and then Ctrl + double LMB will repeat the cut on the exact spot rather than the dot.

I do love this.

https://youtu.be/7SF3yCu53Q0?t=662

Oh right, ther eis this as well. Too bad it is not possible to use dots when moving and scaling.

...Hmmm, maybe it would be possible in multi edit?

https://boxcutter-manual.readthedocs.io/en/latest/hotkeys/

Let me go through the hotkeys. I started those earlier.

4:20pm. https://youtu.be/uCAUpqBK6FE?t=110

Let me try immitating this. I do not understand what he has done here.

4:25pm. https://youtu.be/uCAUpqBK6FE?t=101

I do not get it. What the hell is he doing here?

4:40pm. I am starting to get distracted. Does local mode do something special here.

4:45pm. https://youtu.be/uCAUpqBK6FE?t=28

I got tricked so hard here. It turns out that boxcutter's inset, and Hops inset are not the same thing. The latter applies the booleans (incorrectly) before doing an inset, but the later duplicates the modifiers.

4:50pm. Let me take a break.

5:05pm. Let me resume.

https://youtu.be/uCAUpqBK6FE?t=206

This is cool. I really can't figure out how to fix it in the same way he did.

5:40pm. Insets are a bit convoluted and brittle in the way the work currently.

https://youtu.be/aUd_Ufvfc4U
This is Blender's Best Modeling Tool

This one is by Arrimus.

https://youtu.be/v50cXf6B-2k
boxcutter 718_9: - HOPS Knife Switchblade Off

This is a grand model.

https://youtu.be/v50cXf6B-2k?t=38

Oh, this causes it to reconnect.

6:05pm. Lunch time.

6:25pm. https://youtu.be/4scXty1hsto
Knife Tool Tutorial for Blender 3.0

Done with lunch. Let me watch this.

https://youtu.be/4scXty1hsto?t=82

I had no idea that angle snapping was a thing.

https://youtu.be/4scXty1hsto?t=135

I had no idea it was possible to make multiple cuts like this either.

https://youtu.be/4scXty1hsto?t=196

This really has a lot of functionality I had no idea about.

6:40pm. https://youtu.be/pHpTSLZKIC4?t=320

Now let me go back to this earlier video and I will finish for the day. I feel like I've gotten acclimated with boolean operations in Hops and Boxcutter. I should be able to use it effectively now to do that rig.

https://youtu.be/pHpTSLZKIC4?t=831

I did not know that by pressing P it was possible to adjust the bevel profile.

https://youtu.be/pHpTSLZKIC4?t=1125

Edit multi tool has a join tool.

7:15pm. Figured out what the relative option in inset does.

7:25pm. I am just doing some light mental reviewing. I'll close here for the day. At this point my Blender education should be complete. I'll deal with the rig, the rest of the room and think of what I'll do next with my path.

7:50pm. Maybe if Heaven's Key turns out to be particularly successful that will give me the leeway to cultivate 3d more deeply.

I said I'd try to shot for 4/5 and 5/5 eventually. But honestly this is a huge drag, and it has come in conflict with my desire to actually spend my time productively as opposed to just learning and perfecting my technique.

I am going to shot for 3/5 and not really bother both archmage status in 3d, at least not until I get AI assistants.

Right now, me using 3d to augment my 2d skills, while writing and doing music would be enough to satisfy me.

7:45pm. I feel like I've gone overboard or rather, these topology concerns while working with polygons have taken the wind out of my sails. And topology ends up being interlinked with UV unwrapping and hence texturing. Not to mention shading.

I want to get as far away from that as possible.

I want to ramp up my productivity way up instead of getting stuck pushing vertices around all day like I have for the past half year.

7:50pm. I do not want to dive any deeper into complexity. I want to start getting good at making actual art whether it be 3d or 2d.

7:55pm. I want to let go and just focus the art making. For the past six months it feels like I've just spent my time going through endless tutorials. Who has the will for that?

Enough is enough. I really couldn't avoid having to spend today getting acclimated with Hops and Boxcutter, but from here on out, I will start cranking out the content. I will not give up until I am done.

After I am done with the room, I will loosen the requirements a little and purposely move to a simpler style for the sake of speed. If I can get to an accemptable rate of quality per time, I can consider myself an successful artist."