"10am. Rough night. I went to bed at 2am and had trouble falling asleep. I became obessed with booleaning on curved surfaces. Let me do a bit more on that end.
Before than, let me watch the video by Josh again.

https://youtu.be/qB1eg3ef5vs
Why Your Shading Is Broken (and how to fix it!)

https://youtu.be/qB1eg3ef5vs?t=148

Ah, right. The boolean itself would not make it non-planar, but the bevel would!

10:20am. Focus me. Let me find my peace in booleaning curved surfaces.

11:55am. This is how long it took me to cut in a dozen and a half squares and circles into curved surfaces, clean them up and bevel them. Let me try it in Moi.

I'll just try cutting in a box into a sphere and observing what kind of output comes out.

1:05pm. I am getting filtered massively by 3d here. I spent all this time practicing boolean cleanups on curved surfaces.

This was the first time I tried out Moi 3d's exports. I imported a sphere that had a box cut into it, and to my surprise it has perfect shading!

Meanwhile my own thing is trash and it took 10-100x to create because of all the cleanup.

That other object I was practicing carving into and cleaning up. It is full of distortion when I put it through the striped matcap. It is absolute trash!

It is broken. I really made up my mind to use Hopscutter, but the method is so broken that unless the surface is perfectly flat, I have an enormous amount of cleanup work waiting for me. I even tried doing the subdiv style, but as expected, putting loops everywhere tends to distort the geometry. Working with subdiv is like walking a tightrope. Everything you do affects everything else.

Both the box modeling appoeach with booleans and bevel, and subdiv are just poor trickery.

1:05pm. Yes, I could put in the effort and do the rig using my current skills. If I applied myself, I could definitely do it within a day. But just what is the point in using such obviously broken methods to squeeze out a result? It is not like I am getting paid to make this prop in Blender. Finishing the project, but doing it poorly is no better than not finishing it at all!

1:10pm. I want to get a sense of freedom from art, I do not want to feel like I am doing janitorial work in C.

Blender is out of the running for me for modeling.

I am going to give Zbrush hard surface sculpting a serious try. If I really can't do it there, I'll hack everything out in Moi.

I did not think Moi would have such good shading results. It is really humbling.

1:15pm. Let me get breakfast and then I am going to dedicate myself to HS in Zbrush for a while."