"2pm. Done with breakfast. Let me chill a little. Shit like this just happens. It is in my nature to focus on a particular problem until I deal with it.
The subdiv and bevel styles are obviously wrong to me at this point. Moi might not be obviously right, but it is not obviously wrong.

Sculpting for hard surfaces is something I am going to have to just try.

2:45pm. Now I have a sphere and I am messing with filleting the edges of the cut in box. For the outer rim, masked it out, inverted it and then I can just use the polish deform to fillet it. But I am not really happy with the result. It ends up feeling bloated.

https://youtu.be/jkIf1nNzDKs?t=10
RETOPOLOGY tutorial for HARD SURFACE in Blender

Even more tutorials, let me watch them.

2:55pm. No, this is not the answer. Let me try the hPolish brush.

https://youtu.be/zxc8MvKTqus
Organic/Subd Hard Surface Modeling in Blender (A Brief Overview)

How did he manage this?

https://youtu.be/zxc8MvKTqus?t=548
> I actually didn't even use a bevel modifier. I used the bevel shader.

This is interesting. There is a bevel shader? Is that like the round edge shader for V-Ray?

It is not exactly a surprise that he is not using the bevel mod. It would break everything.

https://youtu.be/zxc8MvKTqus?t=558

Is that in Blender? Yep. Never saw it before.

3:25pm. There must be something wrong with me right now. Remember that course by Cane Towsend. How about I go through the initial part of it.

Right now I am in a state of complete perplexity. I do not even know how to bevel the edges for the cut out properly.

3:55pm. Now I am really focusing on following along what Cane is doing so far so good. Eventually I'll get to the edge work. The part where he explains the tools is not too hard. But I did not have the motivation to follow along last time and it all flew from memory. I only recall what I should be doing, but not how. I know he used the flatten brushes to get the edges.

That seems simple enough, but when I tried it on the circle I could not get it to work for some reason.

4:20pm. This is making me confused. Why is the mask pen brush not working all of a sudden. What did I do?

4:35pm. I don't know. What is happening that the mask is working in reverse mode. I need to switch to the eraser to get it to work normally. Maybe I toggled some setting by accident, but I do not know what.

https://youtu.be/vmCGZk7z0is
AskZBrush: “Why is my Mask Pen brush not masking fully opaque?”

Let me watch this.

4:45pm. Yeah, the pen is inverted. Forge tthis, let me move on. I set the last 2 parts of the course by some other guy to download.

5pm. Let me backtrack a little. How did he max out the mask intensity? By setting the focal shift to -100. Ok.

6pm. Had to take a break. I think I get how he does his thing now. The only thing I do not get exactly is how to do boolean slices without necessarily duplicating the mesh, and doing an intersection + a difference with the cutter.

https://www.reddit.com/r/ZBrush/comments/akgn6g/beginner_question_how_do_you_give_smooth_shading/

It turns Zbrush does not have smooth shading. It instead opts to get that using large poly quantity.

I tried using the Zremesher on the object that I've done, but just like last time it erodes the edges. Unless I want to carry around 100k poly objects even into Clarrise or Blender, I should do my work in Moi.

If the work involves cutting into curved surfaces even a little, I either have to do it in Moi or Zbrush. There is no way around it.

Now, the last part of that course I am downloading will be done in 15m.

Let me rant until then. What I am experiencing here is probably the same as when I did the compilers course in 2015. It took me as long as a week to muster the nerve to even begin work on the code generator for the Cool compiler.

Right now I am in that state. I know a lot of moves, but I am having trouble stepping up to the larger scale than I am used to. So far I've done onyl a simple primitive. The monitor was pretty simple.

Nobody would say that the rig I am attempting is complex, but it is a lot more complex than anything I've attempted previously.

6:05pm. Just a bit more. Just a bit more and I will work up the nerve to do it. If needed I will spend the time time in bed until I find the will.

Now...3m left. Let me just check out the HS techniques from this guy and maybe that will help me come to a conclussion.

6:10pm. Let me take a look at the course.

Udemy - Zbrush Hard Surface Sculpting for All Levels! by Sean Flower

This is the one I got.

> Please try to watch the same tutorial more than once.

Yeah, it is no wonder I've gone back to Cane's course.

6:25pm. Let me just dedicate myself to the Zbrush course. I want more knowledge so I can come to a correct conclussion which approach is best. I'll only feel inner peace once I know the basics properly. This urge to pursue knowledge will go away when I have enough. There is no need to force the matter.

Oh damn it is 10h. I thought it might be smaller since the other course was only 5 despite being twice as big in hard drive space.

This will take a while. I won't take it against me if I take a few days to cover it. I will keep persevering.

7:10pm. Left for lunch. Let me finish the video I am watching and I will call it a day.

This course will go into polygroups, while that radio course barely even touched upon them. It seems that unlike for Blender, Zbrush really needs some pirating in order to learn it. Courses like this one are indispensable.

7:15pm. He is really padding it with the hotkeys. Is the space bar that important? Maybe it is given that the brush stars in the corner instead of the center.

7:25pm. Yeah, the course is most likely 10h long just as it says on the tin. I'll have to speedwatch it before making a decision.

But what is there to decide. I already know that while I do need topology good enough for smooth shading, I do not need it to be sub-divisible. Moi gives me that as well as the ability to boolean on curved surfaces. That means I can just use Blender for blocking out organic shapes and move to Moi when I need to carve them.

Zbrush should not enter into the picture here. But I want it. I want to get just a bit better at this even if I turn my back on it. Maybe later I'll turn my back on 3d modeling as well. For all I know, I could spend the entirety of 2023 doing 2d drawing.

2022 is the time to learn 3d.

7:30pm. This is the way to live. I want to take the time to properly refine my technique. If I can do that, then doing art could be a fun hobby rather than a drudge work."