"10:10am. I am up. Last night I asked on the Moi forums how to do Zbrush style perspective cuts. Before I check out the solution, let me try it on my own. Maybe simply turning off object snapping would do the trick.
Nope. It ends up being drawn on the construction plane which is on the XY axis. I need the contruction plane to be perpendicular to the viewport instead.

https://moi3d.com/forum/messages.php?webtag=MOI&msg=10689.1

///

10689.2 In reply to 10689.1
Hi MRAKGR,

Try setting up a shortcut key to set the construction plane to the view direction of the 3D view:
http://moi3d.com/forum/index.php?webtag=MOI&msg=5759.17
or a different variant here:
http://moi3d.com/forum/index.php?webtag=MOI&msg=10112.11

Then when you trigger that you'll be able to draw on the view oriented construction plane.

To reset it back do a right click on View > CPlane.

- Michael

///

Let me take a look.

10:30am. Hmmm, the cut is sligtly off from where I drew it. I decided to ask about this.

10:35am. Let me chill a bit and I will start. I'll take a break from the Zbrush tutorial and work on the rig in Moi. Consider that a vacation from the tutorials.

11am. Let me start. I had my bit of fun.

First, let me erase the front part and I will redo it.

I simply have to do this. I have more than enough knowledge now. I need to get used to making art.

Now let me make the case 10-20% wider. I hadn't gotten the proportions that wrong, but 10% is just the kind of amount I would not be able to eyeball.

The pen, let me use the pen. After I deal with this, I'll properly set up the hotkeys.

12pm. Sigh, this sometimes happens. Remember when I tried Rhino and it was messing up the fillets. Right now Moi is outright refusing to boolean that sweep that I did. I am not sure what to do here. It would work if I make the profile bigger, but I set it just right.

12:25pm. I spent this entire time trying to figure out how to do that small seam. I did a sweep with the specified profile, but the boolean difference fails for me unless I make it big enough.

Let me make a screenshot and I will ask on the forums. After that I'll leave the issue aside to focus on the other parts. By all means, I could have just done with in Substance.

///

Hi MRAKGR, it's probably because your view is a perspective projection and the cuts are going to project in a straight direction along the construction plane normal.

If you want it to match better you might try setting the 3D view to a parallel projection instead of perspective. You can do that under Options > View > 3D view projection = "Parallel".

- Michael

///

Oh I see. I thought it was parallel already, but I guess I was wrong. I should have guessed it.

12:45pm. Yeah, changing it to parallel (orthogonal) view makes the cuts proper. Come to think of it, I get weird results even in Zbrush unless I switch to orthogonal.

https://moi3d.com/forum/discussion.php?webtag=MOI&msg=10689.5

Here is the thread I opened to ask about the small seam.

Unless I get an answer to this, that will be a point in favor of Zbrush. I am sure than just making the mesh dense enough and then going over it with some brush would have resoved the issue easily enough.

12:50pm. Now let me have breakfast here. Damn it, I hate it when I run into issue like these.

But yeah, I have run into these issues with booleans in the past. Maybe it is because of numerical inaccuracy, but I'd get these weird failures both in Moi and Rhino. I really need to factor these issues in when I decide what my main workflow will be.

Hmph, even an union fails.

https://wiki.mcneel.com/rhino/booleanfaq
> Boolean operations can be great time saving procedures for constructing objects in Rhino. But, for the beginner, they seem to often mysteriously fail, and it is difficult to understand why.

////

So, the first thing to keep in mind is that a solid understanding of how the component operations of Booleans work is necessary. When all else fails, you can get the job done with these commands: Intersect, Split, Delete, and Join. Try undoing a successful Boolean operation and redoing it manually using those commands. Get comfortable with the fact that this procedure will always get you where you want to go even if your Boolean attempts fail.

///

Hmmm, really?

> The other thing to check is your modeling tolerances (see Understanding Tolerances).

This is for Rhino. I'll check out how Moi handles this later. Let have breakfast here."