"2:45pm. Oh, I figured it out. I read something about Rhino not liking when objects are overlapping, so I had the idea to move the curve a little so its ends do not sit directly on the edges. I do not understand why this made it work, but I'll take it.
Let me post the resolution in the thread I opened.

3pm. Done. I am dwelling too much on this. Let me get this party started. Focus me, focus. Last time I got to this point, I lost my nerve. I need to do better this time around.

Got the power button. Let me do the other buttons. What is really important is that things go smoothly for me.

https://youtu.be/XqwMwTEsIcI?t=47

Let me take this chance to plug in these hotkeys. This is the ideal chance to do it.

4:25pm. Things are going well. I've finally hit my stride. Let me see if I can speed things up. I am ungodly slow as usual, but if I can get into the groove, I could put a dent into it.

6:05pm. Done with all the front ports. I'll aim to at least finish the front today so I can leave the easy stuff for tomorrow.

...Time for lunch.

6:20pm. Done with lunch.

6:25pm. I really want to call it a day, but let me do some more.

6:30pm. Let me do it. Just get the thing done.

6:40pm. https://moi3d.com/forum/lmessages.php?webtag=MOI&msg=8123.1
MoI keyboard shortcuts

I'll keep this thread in mind. Let me figure out where the show points is.

7:35pm. Now I've fallen into the trap of hasling myself with a certain edge.

8pm. Oh crap I thought I got it to work right, but I messed it up.

8:45pm. I am finally done with the front part. I think that boolean related issues cost me a few hours today. But I think I finally got my lesson - don't overlap them directly. This is not as easy as it seems, since apparently the whole program is designed to induce overlaps.

8:50pm. Tomorrow, I'll do the last pattern, which should not take me longer than half an hour and after that I'll be done modeling the rig. I'll be able to move to texturing. Right now I am 80% done.

Most of this time has been taken up by the small details.

I can do 95% of the volume in 5% of the time, but then get caught up in the hard stuff that nobody will even notice.

Yeah, this is not the way to really make illustrations fast. If I want to get good at this, I need to stop at blocking out the basic shape. If I want details, I should leave that for the texturing phase.

Hmmm, but yeah. This is not so bad. Boolean issues aside, I'll get used to working around them soon, Moi is a fun little program if you want to do hard surface modeling. Blender despite its complexity, is not better than it.

8:55pm. I'll aim to finish my 3d journey by the end of the month. In June, assuming I am satisfied with how much of the room I've done, I'll move to doing 2d.

I need to follow a process of iterative refinement. I need to get used to sketching in 2d, seeing what the weak points are and then using 3d to strengthen my art. I thought that it might be vible to focus solely on 3d, but it seems there is a reason I've yet to see anybody pursue such a path.

There is no way around it - I am going to have to get good all around at it.

9pm. Assuming I can meet decent quality standard, the most important thing after that is speed.

My sense is that it won't take me longer than six months to really get good at 2d. I've been afraid of testing myself, but my reason is telling me that people who get stuck at beg and int stages have some combination of laziness and stupidity holding them back. I am do not have the first flaw by any stretch surprisingly enough. And when I am stupid, it is due to external circumstance such as boredom, stress or fatigue.

9:05pm. Today's session would have been great if not for these weird boolean issues. But those frustrations aside, I've been in the zone the whole day. I should keep this up.

I can watching tutorials for years, but I'll never break beyond the intermediate stage if that is all that I do. It is when I know the tools, and really practice using them that I step on the path to expertise."