"2:20pm. I thought it would be chilling for longer...let me check out the Birdie Wing thread for just like 5m, and then I will resume.
Ah, fuck. It occured to me to wonder about this just now. It seems that the grate I did and unioned to the main shape is not hollow anymore.

Nevermind that. First let me post the model on the Moi forum and ask them how they would appoach the hole drilling problem.

2:25pm. I am trying to upload the rig, but I keep getting server errors. Nevermind this for now. I'll try posting it later. Right now, let me just redo that grate.

2:40pm. Focus me. Let me break out the pen.

2:50pm. This is ridiculous. Now the network is not acting properly for me.

Let me try uploading the grate error again.

3:15pm. Instead of doing art, I am trying to work around the error cause by nurbs based modeling. One operation would work, the other would mysteriously fail.

3:30pm. Right now I am taking a break to think about it.

4:25pm. Done with it. Holy shit. I newer thought this simple thing would take this long. Forget unioning the grate.

The reason it took so long is because when I modeled the grate I kept running into weird errors with the network node. I can only imagine it was out sync or something.

I am going to count this as a minus in favor of using Moi. Let me just make those screws and I will be done.

https://youtu.be/ewy1WUj-ggc
ZBrush Tutorial - Custom Navigation Middle Mouse Scroll

Let me just watch this.

I really hate using right click to navigate after all. It is really awkward to use the pen like that. I'd rather not use the pen buttons if I can't help it.

5:15pm. Ok, I've learned some important lessons. Next time I'll endeavor not to make things so complicated on myself.

Also I think I understand how the designer create the pattern on the grate. He probably went in and connected the edges menually. I was so damn stubborn with this.

I should not have aimed for founded edges, but chamfers instead. It was not worth it to spend all this time for that little thing. At any rate, I have the grate pattern down including screws.

Let me make the shapes pop just a tiny bit and add the remaining fillets and I will be done.

5:45pm. Finally done! Halleluyah.

This is all I'll need of the model. Let me post it in the wip thread. Then I'll export it and do a distortion check. I'll also check Rhino's quad remesher after that. After that comes texturing.

In hindsight for a semi complicated piece like that grate here which gave me so much trouble in network mode, I should have just subdiv modeled it and exported the control cage into Moi.

When Moi fails it is really surprising and leaves me wondering how to fix it.

5:55pm. Focus me. Let me just post it.

6:05pm. Oh, it managed to union the grate. I am impressed. I wonder why it failed the last time. It is not like I moved it or anything.

7:10pm. Done with lunch. The quad remesher really makes a mess of the grate even with 100k quads just for the big piece. What a waste of time. I won't bother with it again. I think I can safely uninstall Rhino, but I'll keep it around for the time being.

7:15pm. Damn it, let me get a screenshot of the mess.

...It really takes a while to finish.

7:25pm. Finally posted it.

Time to catch up with Fable and UQ Holder.

Tomorrow I will texture the rig. I really should post some of the recent work on Twitter."