"9:55am. Let me read Kaiji and then I will get started with sculpting.
10:15am. Er, let me take a look at the Estab Life thread and then I'll get started. ...No thread. Nwm, let me get started then.

Also I hadn't gotten a reply to that Painter issue yet.

10:25am. Right now I am on vid 10/50 where I left off last time. It seem he is creating a stroke and projecting it all around the sphere. By creating a stroke, it seems he draw it as an outline on the canvas. I've been looking for a feature such as that for a while. Good that I found it.

10:35am. Damn it, the fact that I am starting to lean towards grinding 2d is sapping my will to study this course. I am the kind of guy who if he thinks he has to do something - he will do it. And that includes dropping whatever I am doing to pursue what I should.

Even thought it seems like I have been grinding arbitrary stuff, I've always believed in that the task at hand needed doing.

https://youtu.be/Ll5y29Nfw4k
028 Projection Master Overview

Just what is projection master. The guy doing the course did not explain what he is doing.

Hmmm, ok. Let me give it a try.

http://docs.pixologic.com/user-guide/3d-modeling/sculpting/projection-master/
> Projection Master is a unique feature of ZBrush that allows you to use all of the 2D and 2.5D brushes in sculpting, texturing, and otherwise working with your 3D model.

11:10am. This is so annoying. How do I get out of this stupid thing?

Also for me it only drew on a single side. Let me watch the video again. As well, how did he duplicate the stroke? By pressing S. No, he said S, but he actually pressed Shift + S. He also keeps saying that activating Projection master is G while it is really Shift + G.

What is shift + F? It is doing something, but I am not sure what.

11:25am. I have no idea what is going on anymore. Let me reload the file. I am really on the verge of dropping this course. Let me just finish this video at least.

11:35am. Right now, it works fine for me.

https://youtu.be/I1uZqeefd28
Architectural techniques with Sebastian Legrain using Projection Master - Part 4

This is 13 years old by now, but let me watch it. It seems that nothing is simple when it comes to 3d. It is just tutorials, tutorials, tutorials. Let me finish this and I will move on.

https://youtu.be/I1uZqeefd28?t=76

Yes, if it merely doing those ridges, I could have just used radial symmetry. Why is he bothering me with projection master?

One of the problems that I am having is that I do not understand how to properly dig in.

https://youtu.be/I1uZqeefd28?t=233

I am guessing that at this point they did not have the clip and trim curve brushes.

11:45am. He did not even use projection master.

https://youtu.be/49GN9RtoBkU?t=146

It is really lame how double sides would not work unless you also have the display option turned on.

https://youtu.be/49GN9RtoBkU?t=454

Hmmm, this might have been what was going on.

https://youtu.be/49GN9RtoBkU?t=804

It seems with ZApplink it is possible to set up a canvas in Photoshop and link it to Zbrush via Projection Master.

12:05pm. https://youtu.be/udHde70UZ6A
029 Hard Surface VDM Projection Master

Let me watch this since he will be using it for hard surface shapes.

12:15pm. Let me get back to the course.

12:20pm. Enough, I am done with this.

I do not want to deal with anymore of 3d's complexity.

If I had my mind on a brain core and was able to duplicate myself, I could parallelize the various workloads very neatly in 3d. But right now having to know all this stuff is really killing my pace. 3d is not viable as a 2d replacement. I am absolutely sure of that.

I just do not have the time to develop this path properly as a human.

I am going to sit down and start grinding 2d from here on out. if I need 3d, I'll be sure to bring it in. I can do the rest of the Zbrush course later whenever I feel like it.

12:25pm. Right now I am absoutely sure that nothing will stop me from getting to 3/5 in 3d as long as I put in the effort. But 2d is different. Plenty of people get stuck at the lower levels in it.

I get the sense that I should be able to do it, but I am going to have to make sure of that.

In school I was always bad at it, but I don't think I've ever had the sense of breaking the scene down into the constituent steps. I never even understood the purpose of sketching back then, so it is no wonder I was so poor at it.

Every time I did art back then, I had the image of what I wanted to do and just started mucking about to poor results. I just dove in without planning and got discouraged when it did not give me the results that I wanted.

I don't think I've ever improved despite years of art classes in school.

12:30pm. I am simply going to have to face this demon and hunt it down.

It is one thing to draw or paint from imagination, but I should get to a high level in copying what is in front of me. There is just no reason I should ever be weak at that as an artist.

12:35pm. I am going to make my first target the middle and back of the room. The monitor, the rig, the desk and the radiator are near the entrance at the back, but turning 90 degrees would give me the beds and the balcony door as well as the shelves and the boxes on top with the rig being partially to the side.

12:35pm. I am just so tired at this point.

Pumping out elaborate 3d scenes is beyond my ability unless I want to take a month per each. I might get faster if I persevere, but I'll never be able to overcome the strain of constant context switching between several different applications.

With Clip Studio Paint, I have a single tool that could do anything I'd need. Just that would speed things up significantly.

Right, remember months ago when I was studying Blender and it turned out some of the scenes that were very elaborate were just textures on various planes? I am going to have to investigate that kind of 2.5d workflow. I'll leave that for later, it is not like I need to get into that in particular right now. I know more than enough Blender to know how to immitate such a style already.

12:40pm. I want to get better.

Forget great things, I just need to get good at painting what is in front of me. If I can copy what I see, that will form the foundation for everything else.

Learning 3d was a great exercise in learning how to break things down into their primitives. That is what I need to focus on if I want to get good at 2d.

In school I was not aware of how a painting can be gradually refined. But now I definitely understand that.

It might turn out to be that I've been supressing myself too much with my timidity towards just drawing on a canvas.

12:45pm. If I really can't do it and am one of those retards who can't move beyond the /beg/ stage, I'll go back to programming and get a job. If I can't do art, then I'll have to resolve myself to become a wage slave."